refactor model classes - move some classes to base package - update package name in other class that using them
