+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/RemoteCalendarEvents.kt
@@ -2,7 +2,7 @@ package com.github.persiancal.sdkremote
 
 import android.content.Context
 import android.util.Log
-import com.github.persiancal.sdkremote.model.jalali.JalaliResponse
+import com.github.persiancal.sdkremote.model.base.JalaliResponse
 import com.github.persiancal.sdkremote.model.jalali.MyObjectBox
 import com.github.persiancal.sdkremote.model.jalali.RemoteJalaliEventsDb
 import com.github.persiancal.sdkremote.model.jalali.RemoteJalaliEventsDb_
