+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/model/base/Months.kt
@@ -1,4 +1,4 @@
-package com.github.persiancal.sdkremote.model.jalali
+package com.github.persiancal.sdkremote.model.base
 
 import com.google.gson.annotations.SerializedName
