+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/model/base/NameItem.kt
@@ -1,8 +1,8 @@
-package com.github.persiancal.sdkremote.model.jalali
+package com.github.persiancal.sdkremote.model.base
 
 import com.google.gson.annotations.SerializedName
 
-data class CalendarsItem(
+data class NameItem(
 
     @field:SerializedName("en_US")
     val enUS: String? = null,
